For dumb people like me give ready to copy/paste commands
I observed some odd behavior when running the demo and it is because I
changed to port number for the tui but did not notice I also had to
change the key files.

That's my bad but I think with this form, we reduce the risk of mistake
and we make it more coherent with the previous section.