Rationalize doc step
We do not optimize docusaurus build when not on master or release.
We only want to sanity chack the doc and the links but do not plan
to ever publish this doc. Not optimizing saves minutes.

In case we're building for master or release, then, we optimize.

Also, we did not need both stable/unstable versions for master so
we only build the doc once.